requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiogram_dialog==2.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
environs==11.0.0
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.1.0
idna==3.8
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.1.0
openpyxl==3.1.5
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.13.1